move graph methods from tcp into actions
